+++ console2/src/components/organisms/CheckpointView/CheckpointViewContainer/loadData.ts
@@ -28,7 +28,10 @@ export interface FetchProcessArgs {
     offset?: number;
 }
 
-export const loadData = (args: FetchProcessArgs) => async ({ setState, state }: any): Promise<ProcessEntry[] | undefined> => {
+export const loadData = (args: FetchProcessArgs) => async ({
+    setState,
+    state
+}: any): Promise<ProcessEntry[] | undefined> => {
     if (state.loadingData) {
         return;
     }
@@ -39,7 +42,10 @@ export const loadData = (args: FetchProcessArgs) => async ({ setState, state }: 
     const checkpointGroups = {};
     processes.forEach((p) => {
         if (p.checkpoints && p.statusHistory) {
-            checkpointGroups[p.instanceId] = generateCheckpointGroups(p.checkpoints, p.statusHistory);
+            checkpointGroups[p.instanceId] = generateCheckpointGroups(
+                p.checkpoints,
+                p.statusHistory
+            );
         }
     });