+++ server/impl/src/main/java/com/walmartlabs/concord/server/ServerModule.java
@@ -26,6 +26,7 @@ import com.walmartlabs.concord.db.DatabaseModule;
 import com.walmartlabs.concord.server.agent.AgentCommandWebSocketHandler;
 import com.walmartlabs.concord.server.metrics.MetricModule;
 import com.walmartlabs.concord.server.process.queue.ProcessQueueWebSocketHandler;
+import com.walmartlabs.concord.server.security.rememberme.RememberMeCookieCleaner;
 import com.walmartlabs.concord.server.task.TaskScheduler;
 
 public class ServerModule extends AbstractModule {
@@ -39,5 +40,6 @@ public class ServerModule extends AbstractModule {
         tasks.addBinding().to(AgentCommandWebSocketHandler.class);
         tasks.addBinding().to(ProcessQueueWebSocketHandler.class);
         tasks.addBinding().to(TaskScheduler.class);
+        tasks.addBinding().to(RememberMeCookieCleaner.class);
     }
 }