+++ server/impl/src/main/java/com/walmartlabs/concord/server/security/PrincipalUtils.java
@@ -9,9 +9,9 @@ package com.walmartlabs.concord.server.security;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -24,6 +24,9 @@ import org.apache.shiro.SecurityUtils;
 import org.apache.shiro.authc.AuthenticationException;
 import org.apache.shiro.subject.PrincipalCollection;
 import org.apache.shiro.subject.Subject;
+
+import java.io.*;
+import java.util.Optional;
 
 public final class PrincipalUtils {
 
@@ -47,6 +50,29 @@ public final class PrincipalUtils {
             throw new AuthenticationException("Can't determine the current principal (" + type.getName() + ")");
         }
         return p;
+    }
+
+    public static byte[] serialize(PrincipalCollection data) {
+        ByteArrayOutputStream baos = new ByteArrayOutputStream();
+        try (ObjectOutputStream oos = new ObjectOutputStream(baos)) {
+            oos.writeObject(data);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+        return baos.toByteArray();
+    }
+
+    public static Optional<PrincipalCollection> deserialize(byte[] data) {
+        InputStream in = new ByteArrayInputStream(data);
+        return deserialize(in);
+    }
+
+    public static Optional<PrincipalCollection> deserialize(InputStream in) {
+        try (ObjectInputStream ois = new ObjectInputStream(in)) {
+            return Optional.of((PrincipalCollection) ois.readObject());
+        } catch (IOException | ClassNotFoundException e) {
+            throw new RuntimeException(e);
+        }
     }
 
     private PrincipalUtils() {